% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zika}
\alias{zika}
\title{Zika reports}
\format{A data frame with the following variables:
\itemize{
  \item report_date: Date of the report
  \item location: Location
}}
\usage{
zika

zika
}
\description{
Curated data set from this GitHub repo -- \url{https://github.com/cdcepi/zika}

Curated data set from this GitHub repo -- \url{https://github.com/cdcepi/zika}
}
\keyword{datasets}

